requirements.txt
python>=3.6
pytorch>=1.1.0
tensorboard>=1.14.0
joblib>=0.14.1